refactor: 로그아웃 버튼 위치 변경
